+++ app/src/test/java/javaroke/queue/QueueAbstractTest.java
@@ -1,9 +1,8 @@
-package javaroke;
+package javaroke.queue;
 
 import org.junit.Test;
 import static org.junit.Assert.*;
 
-import javaroke.queue.QueueAbstract;
 
 public class QueueAbstractTest {
   @Test
